reorganize the code of RawYm2612
