+++ CMakeLists.txt
@@ -82,7 +82,8 @@ set(FILEFORMATS_SOURCES
   "src/FileFormats/format_opm.cpp"
   "src/FileFormats/format_mucom88_dat.cpp"
   "src/FileFormats/format_vgi.cpp"
-  "src/FileFormats/format_wohlstand_opn2.cpp")
+  "src/FileFormats/format_wohlstand_opn2.cpp"
+  "src/FileFormats/ym2612_to_wopi.cpp")
 add_library(FileFormats STATIC ${FILEFORMATS_SOURCES})
 target_include_directories(FileFormats PUBLIC "src")
 target_link_libraries(FileFormats PUBLIC Common)
