+++ FMBankEdit.pro
@@ -119,6 +119,7 @@ SOURCES += \
     src/FileFormats/format_mucom88_dat.cpp \
     src/FileFormats/format_vgi.cpp \
     src/FileFormats/format_wohlstand_opn2.cpp \
+    src/FileFormats/ym2612_to_wopi.cpp \
     src/formats_sup.cpp \
     src/importer.cpp \
     src/latency.cpp \
@@ -155,6 +156,7 @@ HEADERS += \
     src/FileFormats/format_mucom88_dat.h \
     src/FileFormats/format_vgi.h \
     src/FileFormats/format_wohlstand_opn2.h \
+    src/FileFormats/ym2612_to_wopi.h \
     src/formats_sup.h \
     src/importer.h \
     src/latency.h \
@@ -168,8 +170,7 @@ HEADERS += \
     src/opl/realtime/ring_buffer.h \
     src/opl/realtime/ring_buffer.tcc \
     src/piano.h \
-    src/version.h \
-    src/FileFormats/ym2612_to_wopi.hpp
+    src/version.h
 
 FORMS += \
     src/bank_editor.ui \
